+++ tests/test_client.py
@@ -42,7 +42,6 @@ from cscapi.client import (
     CAPI_WATCHER_REGISTER_ENDPOINT,
     CAPI_METRICS_ENDPOINT,
     CAPIClient,
-    has_valid_token,
     CAPIClientConfig,
 )
 from cscapi.sql_storage import SQLStorage
@@ -195,7 +194,7 @@ class TestChooseEnv:
 
 class TestSendSignals:
     def test_fresh_send_signals(self, httpx_mock: HTTPXMock, client: CAPIClient):
-        assert len(client.storage.get_all_signals()) == 0
+        assert len(client.storage.get_all_signals(limit=1000)) == 0
         token = dummy_token()
         httpx_mock.add_response(
             method="POST",
@@ -217,7 +216,7 @@ class TestSendSignals:
         s1 = replace(mock_signals()[0], scenario="crowdsecurity/http-bf")
         s2 = mock_signals()[0]
         client.add_signals([s1, s2])
-        assert len(client.storage.get_all_signals()) == 2
+        assert len(client.storage.get_all_signals(limit=1000)) == 2
 
         assert client.storage.get_machine_by_id("test") is None
 
@@ -247,7 +246,7 @@ class TestSendSignals:
     def test_signal_gets_deleted_after_send(
         self, httpx_mock: HTTPXMock, client: CAPIClient
     ):
-        assert len(client.storage.get_all_signals()) == 0
+        assert len(client.storage.get_all_signals(limit=1000)) == 0
         token = dummy_token()
         httpx_mock.add_response(
             method="POST",
@@ -268,9 +267,9 @@ class TestSendSignals:
 
         s1 = mock_signals()[0]
         client.add_signals([s1])
-        assert len(client.storage.get_all_signals()) == 1
+        assert len(client.storage.get_all_signals(limit=1000)) == 1
         client.send_signals(prune_after_send=True)
-        assert len(client.storage.get_all_signals()) == 0
+        assert len(client.storage.get_all_signals(limit=1000)) == 0
 
     def test_signals_from_already_registered_machine(
         self, httpx_mock: HTTPXMock, client: CAPIClient